Refactor Block class into geometry module
