+++ qlbm/lattice/geometry/shapes/__init__.py
@@ -1,0 +1,5 @@
+"""Geometrical shapes for solid geometry."""
+
+from .block import Block
+
+__all__ = ["Block"]
