+++ qlbm/lattice/lattices/collisionless_lattice.py
@@ -6,7 +6,7 @@ from typing import Dict, List, Tuple
 from qiskit import QuantumCircuit, QuantumRegister
 from typing_extensions import override
 
-from qlbm.lattice.blocks import Block
+from qlbm.lattice.geometry.shapes.block import Block
 from qlbm.tools.exceptions import LatticeException
 from qlbm.tools.utils import dimension_letter, flatten
